Moved tests to UnitTest project
