+++ MBINCompiler.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
-MinimumVisualStudioVersion = 10.0.40219.1
+# Visual Studio 2012
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MBINCompiler", "MBINCompiler\MBINCompiler.csproj", "{CE736EF9-7BE9-4FF0-9551-78F0071AA373}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MBINCompilerTests", "MBINCompilerTests\MBINCompilerTests.csproj", "{836B3A35-C1EA-47C8-B3AB-59C0E2E0C763}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +15,10 @@ Global
 		{CE736EF9-7BE9-4FF0-9551-78F0071AA373}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CE736EF9-7BE9-4FF0-9551-78F0071AA373}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CE736EF9-7BE9-4FF0-9551-78F0071AA373}.Release|Any CPU.Build.0 = Release|Any CPU
+		{836B3A35-C1EA-47C8-B3AB-59C0E2E0C763}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{836B3A35-C1EA-47C8-B3AB-59C0E2E0C763}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{836B3A35-C1EA-47C8-B3AB-59C0E2E0C763}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{836B3A35-C1EA-47C8-B3AB-59C0E2E0C763}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
