+++ MBINCompiler/Program.cs
@@ -24,44 +24,6 @@ namespace MBINCompiler
         // todo: align based on offset of the struct (not offset of the reader!)
         static void Main(string[] args)
         {
-            //List<string> types = new List<string>();
-            //ScanMBINs("C:\\NMS\\", ref types);
-            //File.WriteAllLines("C:\\NMS\\mbins.txt", types.ToArray());
-
-            var tests = new string[] // all of these should decompile fine
-            {
-                @"C:\NMS\METADATA\UI\HUD\SHIPHUD.MBIN", 
-                @"C:\NMS\METADATA\ENTITLEMENTS\GCENTITLEMENTREWARDTABLE.MBIN",
-                @"C:\NMS\SCENES\DEMOS\PS4TEST\MAINSETTINGSPS4TEST.MBIN",
-                @"C:\NMS\METADATA\REALITY\TABLES\NMS_REALITY_GCTECHNOLOGYTABLE.MBIN",
-                @"C:\NMS\METADATA\SIMULATION\SOLARSYSTEM\WEATHER\CLEARWEATHER.MBIN",
-                @"C:\NMS\METADATA\SIMULATION\SOLARSYSTEM\WEATHER\SKYSETTINGS\DAYSKYCOLOURS.MBIN",
-                @"C:\NMS\LANGUAGE\NMS_LOC1_ENGLISH.MBIN",
-                @"C:\NMS\METADATA\REALITY\TABLES\NMS_DIALOG_GCALIENSPEECHTABLE.MBIN",
-                @"C:\NMS\METADATA\REALITY\TABLES\NMS_DIALOG_GCALIENPUZZLETABLE.MBIN"
-            };
-
-            var penaltyBox = new string[]
-            {
-            };
-
-            foreach(var test in tests)
-            {
-                var file2 = new MBINFile(test);
-                file2.Load();
-                var wew = file2.SerializeToXML();
-                wew = wew;
-            }
-
-            /*
-            foreach (var test in penaltyBox)
-            {
-                var file2 = new MBINFile(test);
-                file2.Load();
-                var wew = file2.SerializeToXML();
-                wew = wew;
-            }*/
-
             if (args.Length < 1)
             {
                 Console.WriteLine("Usage: MBINCompiler [InputPath]");
@@ -71,6 +33,7 @@ namespace MBINCompiler
             }
 
             var input = args[0];
+
             if(Path.GetExtension(input).ToLower() != ".mbin")
             {
                 Console.WriteLine("Unsupported input file extension, only .mbin is currently supported, exiting...");
